requirements.txt
opencv-contrib-python>=4.8.0
volcengine-python-sdk[ark]
python-dotenv
PySide6>=6.5.0
numpy
pyyaml
Pillow>=9.0.0
